Now storing worker command in result db.
Workers now report the actual command they executed (e.g. cosmic-ray
worker . . .) in their results. We store this in the results database.

I'm doing this right now because these commands are useful for
debugging. In the long run it may not be a great idea...we can just
strip it out at that point.