Introduce new parameter `num_mutations`
it is used for operators which can produce multiple mutations
for the same piece of code, for example `A and B and C` can become
`A or B and C` or `A and B or C` or `A or B or C`! The num_mutations
parameter is used by MutatingCore.visit_mutation_site() and passed
as an index to Operator.mutate(). mutate() will use the index to decide
which of all possible mutation variants to produce.