Refactor MutatingCore.visit_mutation_site(), fixes #174
Explanation from @abingham:

When operators were only allowed to make one mutation per node,
self._target represented the N-th time an operator encountered a
node that it could mutate. Now self._target represents the N-th
time an operator could perform a mutation, taking into account that
an operator might make multiple mutations per node.

Similarly, self._count previously represented how many mutatable nodes
had been visited; now it represents how many mutations an operator has
so far claimed it could make while walking the AST.

So the basic idea of the change is to say this: if the operator claims
that it performs enough mutations on the current node to reach the
target, then let it do the mutation. The idx that we pass back to the
operator is the span value, i.e. the distance from the current
_count to the _target.