pytest: move config to setup.cfg, only use testpaths
There is no need for `norecursedirs` anymore.

Having the config in a more central location simplifies it, and avoids
having to pass in `test` via posargs in tox.

This also moves the flake8 cfg to setup.cfg, keeping tox.ini just for
tox, while setup.cfg is used for generic tool config.