Issue #565 - Avoid mutating pipe operator in assignment annotations
Currently, mutating pipes in the annotations of assignments create false mutant survivals because there is not runtime implications (errors or otherwise).

This patch will prevent the pipe character within the annotations of assignments from being seen as a binary operator, thus preventing mutations of that character and eliminating the false-positive mutant survival.