Potential fix to `mutate-and-test` command
While in documentation of mutate-and-test command it says that it must be passed the module path, operator, occurrence and test_command, when one send a command like that, an error was presented to the user like shown below, saying that 3 positional arguments were required and not five.

The apparent reason was the mismatch in the arguments passed to `mutate_and_test`: module path, operator and ocurrence are passed as a MutationSpec. I took the typing example from test suite and so far I am unsure about the value of start and end position. Is there a function to get that dynamically?

Documentation usage:

Usage: cosmic-ray mutate-and-test [OPTIONS] MODULE_PATH OPERATOR OCCURRENCE
                                  TEST_COMMAND

Sample command:

```
cosmic-ray mutate-and-test {source_file} core/ReplaceBinaryOperator_Add_Sub 0 "pytest -r A --code-highlight=yes -v"
```

Error message in version 8.4.3:

```
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "\.venv\Scripts\cosmic-ray.exe\__main__.py", line 7, in <module>
    sys.exit(main())
             ~~~~^^
  File "\.venv\Lib\site-packages\cosmic_ray\cli.py", line 293, in main
    return cli(argv)
  File "\.venv\Lib\site-packages\click\core.py", line 1462, in __call__
    return self.main(*args, **kwargs)
           ~~~~~~~~~^^^^^^^^^^^^^^^^^
  File "\.venv\Lib\site-packages\click\core.py", line 1383, in main
    rv = self.invoke(ctx)
  File "\.venv\Lib\site-packages\click\core.py", line 1850, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
                           ~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^
  File "\.venv\Lib\site-packages\click\core.py", line 1246, in invoke
    return ctx.invoke(self.callback, **ctx.params)
           ~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "\.venv\Lib\site-packages\click\core.py", line 814, in invoke
    return callback(*args, **kwargs)
  File "\.venv\Lib\site-packages\cosmic_ray\cli.py", line 270, in mutate_and_test
    work_result = cosmic_ray.mutating.mutate_and_test(
        Path(module_path), operator, occurrence, test_command, None
    )
TypeError: mutate_and_test() takes 3 positional arguments but 5 were given
```